Handle multi-line if expressions better
Exclude a line from being treated as a continuing expression if it
starts with `else`. I'm not sure where else in the grammar this would
show up.

Also added quotes around ${EMACS} in run_tests.sh to make it easier to
run the tests on windows (my Emacs is under "C:\Program Files").