Replace obsolete function `point-at-eol` with equivalent.
Emacs 29.1+ emits a warning when code using `point-at-eol` is compiled; see
emacs/master commit b7e867. This is a trivial but safe change to get rid of that
warning. `point-at-eol` seems to have been the same thing as `line-end-position`
since 1999 (emacs/master commit 475fb2f). I've tested this briefly and the
behavior of `zig-end-of-defun` does seem to be the same with this change.